Create separate currency module with class refinements, break separate reports into files, style fixes
