+++ bin/shopify_dashboard_plus.rb
@@ -1,4 +1,5 @@
 #!/usr/bin/env ruby
+# frozen_string_literal: true
 
 require File.expand_path('../../lib/shopify_dashboard_plus.rb', __FILE__)
 require 'vegas'
