+++ config.ru
@@ -1,2 +1,4 @@
+# frozen_string_literal: true
+
 require './lib/shopify_dashboard_plus.rb'
 run Sinatra::Application
