+++ test/resources/modify_data.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 module ModifyData
   Faker::Config.locale = :"en-CA"
