+++ test/test_mockdata.rb
@@ -9,7 +9,10 @@ require 'byebug'
 require './lib/shopify_dashboard_plus.rb'
 require './lib/shopify_dashboard_plus/version'
 require './lib/shopify_dashboard_plus/helpers'
-require './lib/shopify_dashboard_plus/report'
+require './lib/shopify_dashboard_plus/discount_report'
+require './lib/shopify_dashboard_plus/revenue_report'
+require './lib/shopify_dashboard_plus/sales_report'
+require './lib/shopify_dashboard_plus/traffic_report'
 
 ENV['RACK_ENV'] = 'test'
